requirements.txt
Django==2.1
djangorestframework==3.8.2
Django-environ==0.4.5
factory_boy==2.11.1
